initial reordering
